split index.ts file
